requirements.txt
Django==1.9.8
django-taggit==0.20.2
pytz==2016.6.1
